pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "priv-goals"
version = "0.0.1"
description = "A privacy-focused goal tracking application powered by A.I."
readme = "README.md"
authors = [{ name = "Really Him", email = "hesreallyhim@proton.me" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["goals", "productivity", "ai", "privacy"]
dependencies = [
    "gradio>=4.0.0",
    "litellm>=1.0.0",
    "python-dotenv>=0.19.0",
    "gspread>=5.0.0",
    "oauth2client>=4.1.3",
    "inquirer>=3.1.0",
    "pyyaml>=6.0.0",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = ["black", "pylint", "pytest"]

[project.scripts]
priv-goals = "priv_goals.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["priv_goals"]
